tests/wpt/tests/css/css-shadow-parts/complex-non-matching.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Shadow Parts - Complex Non-matching</title>
    <meta href="mailto:fergal@chromium.org" rel="author" title="Fergal Daly">
    <link href="http://www.google.com/" rel="author" title="Google">
    <link href="https://drafts.csswg.org/css-shadow-parts/" rel="help">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/shadow-helper.js"></script>
  </head>
  <body>
    <style>div #c-e::part(partp) { color: red; }</style>
    <script>installCustomElement("custom-element", "custom-element-template");</script>
    <template id="custom-element-template">
      <style>span { color: green; }</style>
      <span id="part" part="partp">This text</span>
    </template>
    The following text should be green:
    <pre><custom-element id="c-e"></custom-element></pre>
    <script>
      "use strict";
      const colorGreen = "rgb(0, 128, 0)";
      test(function() {
        const el = getElementByShadowIds(document, ["c-e", "part"]);
        assert_equals(window.getComputedStyle(el).color, colorGreen);
      }, "Non-matching complex selector for host does not style");
    </script>
  </body>
</html>
